+++ .gitignore
@@ -17,3 +17,4 @@
 *.app
 
 test-fs
+tmp/1
