+++ Makefile
@@ -1,11 +1,12 @@
+CFLAGS?=--std=c99
 
 all: clean test
 
 clean:
-	rm -f test-fs
+	rm -f test-fs *.o
 
-test:
-	$(CC) --std=c99 test.c -o test-fs
+test: fs.o
+	$(CC) ${CFLAGS} test.c fs.o -o test-fs
 	./test-fs
 
 .PHONY: clean test
